# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: RESTful APIs of Material Management App
host: localhost:3000
basePath: /api
schemes:
  - http
  - https
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: |
      For accessing the API a valid JWT token must be passed in all the queries in
      the 'Authorization' header.


      A valid JWT token is generated by the API and retourned as answer of a call
      to the route /login giving a valid user & password.


      The following syntax must be used in the 'Authorization' header :

          Bearer xxxxxx.yyyyyyy.zzzzzz
consumes:
  - application/json
produces:
  - application/json
paths:
  /oauth/token:
    x-swagger-router-controller: main-controller
    post:
      tags:
        - User Authentication
      operationId: loginPost
      summary: "Authorization for the access of other APIs"
      description: "Login with a valid credential"
      parameters:
        - name: authentication
          in: body
          required: true
          schema:
            $ref: "#/definitions/Authentication"
      responses:
        "200":
          description: "Success"
          schema:
            $ref: "#/definitions/Token"
        "404":
          description: "Access Denied"
          schema:
            $ref: "#/definitions/Error"
        "500":
          description: "Invalid Access Token!!"
          schema:
            $ref: "#/definitions/Error"
  /appUser/{user_id}:
    x-swagger-router-controller: main-controller
    get:
      tags:
        - User Authentication
      operationId: appUser
      summary: "Get App User Involvements"
      description: "Fetch app-user associated with the authorized user"
      # security:
      #   - Bearer: []
      x-security-scopes:
        - admin
      parameters:
        - name: "user_id"
          in: "path"
          description: "User ID"
          required: true
          type: "string"
      responses:
        "200":
          description: "Success"
          schema:
            type: array
            items:
              $ref: "#/definitions/AppUser"
        "404":
          description: "Access Denied"
          schema:
            $ref: "#/definitions/Error"
...
  /projectByDepartment/{department_key}:
    x-swagger-router-controller: main-controller
    get:
      tags:
        - Manage & View Requests (Backfill/Disposal)
      operationId: projectByDepartment
      summary: "List all projects by a given depertment key"
      description: "Get details of all projects by a selected department"
      # security:
      #   - Bearer: []
      x-security-scopes:
        - admin
      parameters:
        - name: "department_key"
          in: "path"
          description: "Department Key"
          required: true
          type: "string"
      responses:
        "200":
          description: "Success"
          schema:
            $ref: "#/definitions/Project"
        "404":
          description: "Access Denied"
          schema:
            $ref: "#/definitions/Error"
...
  /listUserRequests/{user_id}:
    x-swagger-router-controller: list_requests-controller
    get:
      tags:
        - Manage & View Requests (Backfill/Disposal)
      operationId: listUserRequests
      summary: "List User Created Disposal/Backfill Requests"
      description: "Get User Created Disposal/Backfill Requests"
      # security:
      #   - Bearer: []
      x-security-scopes:
        - admin
      parameters:
        - name: "user_id"
          in: "path"
          description: "User ID"
          required: true
          type: "string"
        # - name: "request_type"
        #   in: "path"
        #   description: "Request Type"
        #   required: true
        #   type: "string"
        #   enum: [Disposal, Backfill]
      responses:
        "200":
          description: "Success"
          schema:
            type: array
            items:
              $ref: "#/definitions/ListRequest"
        "404":
          description: "Access Denied"
          schema:
            $ref: "#/definitions/Error"
  /listGlobalRequests/{user_id}/{request_type}:
    x-swagger-router-controller: list_requests-controller
    get:
      tags:
        - Manage & View Requests (Backfill/Disposal)
      operationId: listGlobalRequests
      summary: "List Global Disposal/Backfill Requests"
      description: "Get Disposal/Backfill Requests which are not Created by User"
      # security:
      #   - Bearer: []
      x-security-scopes:
        - admin
      parameters:
        - name: "user_id"
          in: "path"
          description: "User ID"
          required: true
          type: "string"
        - name: "request_type"
          in: "path"
          description: "Request Type"
          required: true
          type: "string"
          enum: [Disposal, Backfill]
      responses:
        "200":
          description: "Success"
          schema:
            type: array
            items:
              $ref: "#/definitions/ListRequest"
        "404":
          description: "Access Denied"
          schema:
            $ref: "#/definitions/Error"
...
  /projects/{id}:
    x-swagger-router-controller: project-controller
    get:
      tags:
        - Projects
      operationId: listProjects
      summary: "Get projects"
      description: "Fetch app projects"
      # security:
      #   - Bearer: []
      x-security-scopes:
        - admin
      parameters:
        - name: "id"
          in: "path"
          description: "ID"
          required: true
          type: "string"
      responses:
        "200":
          description: "Success"
          schema:
            type: array
            items:
              $ref: "#/definitions/AppUser"
        "404":
          description: "Access Denied"
          schema:
            $ref: "#/definitions/Error"
  /projectStaff/{department_key}:
    x-swagger-router-controller: project-controller
    get:
      tags:
        - Projects
      operationId: projectStaff
      summary: "Get project staff"
      description: "Fetch app project staff"
      # security:
      #   - Bearer: []
      x-security-scopes:
        - admin
      parameters:
        - name: "department_key"
          in: "path"
          description: "ID"
          required: true
          type: "string"
      responses:
        "200":
          description: "Success"
          schema:
            type: array
            items:
              $ref: "#/definitions/AppUser"
        "404":
          description: "Access Denied"
          schema:
            $ref: "#/definitions/Error"
  /allProjectByPage/{currentPage}/{pageSize}:
    x-swagger-router-controller: project-controller
    get:
      tags:
        - Projects
      operationId: allProjectByPage
      summary: "Get projects by page"
      description: "Fetch app projects by page"
      # security:
      #   - Bearer: []
      x-security-scopes:
        - admin
      parameters:
        - name: "currentPage"
          in: "path"
          description: "Page number"
          required: true
          type: "string"
        - name: "pageSize"
          in: "path"
          description: "Page size"
          required: true
          type: "string"
      responses:
        "200":
          description: "Success"
          schema:
            type: array
            items:
              $ref: "#/definitions/AppUser"
        "404":
          description: "Access Denied"
          schema:
            $ref: "#/definitions/Error"
definitions:
  Authentication:
    type: object
    properties:
      email:
        type: string
        example: "user2@hiphing.com.hk"
      password:
        type: string
        example: "pw2"
  AppUser:
    properties:
      user_id:
        type: "string"
      staff_id:
        type: "string"
      last_name:
        type: "string"
      first_name:
        type: "string"
      english_name:
        type: "string"
      chinese_name:
        type: "string"
      email:
        type: "string"
      department_key:
        type: "string"
      department_id:
        type: "string"
      dept_name:
        type: "string"
      dept_chi_name:
        type: "string"
      position_key:
        type: "string"
      position_name:
        type: "string"
      post_name:
        type: "string"
      post_chi_name:
        type: "string"
      role_id:
        type: "string"
...
  Project:
    properties:
      department_key:
        type: "string"
      department_id:
        type: "string"
      projectName:
        type: "string"
      projectNameChin:
        type: "string"
      projDescription:
        type: "string"
      address:
        type: "string"
      govtRefNo:
        type: "string"
      projStatus:
        type: "string"
      rstatus:
        type: "string"
      latitude:
        type: "string"
      longitude:
        type: "string"
      contractType:
        type: "string"
  ListRequest:
    properties:
      id:
        type: "string"
      department_key:
        type: "string"
      filling_purpose:
        type: "string"
        enum: [Permanent, Temporary]
      material_type:
        type: "string"
      material_quality:
        type: "string"
      material_volume:
        type: "number"
        multipleOf: 0.01
      schedule_start_date:
        type: "string"
        format: "date-time"
      schedule_end_date:
        type: "string"
        format: "date-time"
...
  Token:
    properties:
      access_token:
        type: "string"
      token_type:
        type: "string"
      expires_in:
        type: "integer"
        format: "int64"
      user_id:
        type: "string"
      scope:
        type: "string"
...
  Error:
    type: object
    properties:
      message:
        type: string